+++ package.json
@@ -34,7 +34,7 @@
 		"ts-node": "^10.9.1",
 		"ts-toolbelt": "^9.6.0",
 		"tsconfig-paths": "^4.1.2",
-		"typescript": "^4.9.5"
+      "typescript": "^5.4.5"
 	},
 	"dependencies": {
 		"@discordjs/builders": "^1.5.0-dev.1677499471-ffdb197.0",
