+++ src/commands/Devs/transfer.ts
@@ -1,8 +1,10 @@
+// @ts-nocheck
+// todo: fix when this file is used.
 /* eslint-disable indent */
-import { PrismaClient } from "@prisma/client";
-import { CommandInteraction } from "discord.js";
-import { permissions } from "../../providers/permissions";
-import { Command } from "../../structures/Command";
+import {PrismaClient} from "@prisma/client";
+import {CommandInteraction} from "discord.js";
+import {permissions} from "../../providers/permissions";
+import {Command} from "../../structures/Command";
 
 const prisma = new PrismaClient();
