+++ src/commands/economy/daily.ts
@@ -1,11 +1,12 @@
 //daily.ts
-import { db } from "../../database/database";
-import { upsertUserInfo } from "../../database/userInfo";
-import { constants, text } from "../../providers/config";
-import { Command } from "../../structures/Command";
-import { format } from "../../utils/string";
+import {db} from "../../database/database";
+import {upsertUserInfo} from "../../database/userInfo";
+import {constants, text} from "../../providers/config";
+import {Command} from "../../structures/Command";
+import {format} from "../../utils/string";
 import pms from "pretty-ms";
-import { randRange, sampleArray } from "../../utils/utils";
+import {randRange, sampleArray} from "../../utils/utils";
+
 const cooldowns: Record<string, number> = {};
 
 export const command = new Command("daily", "Get your daily income!.")
@@ -19,7 +20,7 @@ export const command = new Command("daily", "Get your daily income!.")
 			);
 			return;
 		}
-		const info = await upsertUserInfo(int.user, int.guild?.id || '');
+		const info = await upsertUserInfo(int.user);
 		const obtained = randRange(...constants.daily.amountRange);
 		cooldowns[int.user.id] = Date.now() + constants.daily.cooldownMs;
 		await db.userInfo.update({ where: { id: info.id }, data: { balance: { increment: obtained } } });
