+++ src/commands/economy/work.ts
@@ -1,13 +1,14 @@
 /* eslint-disable quotes */
 /* eslint-disable linebreak-style */
 //work.ts
-import { db } from "../../database/database";
-import { upsertUserInfo } from "../../database/userInfo";
-import { constants, text } from "../../providers/config";
-import { Command } from "../../structures/Command";
-import { format } from "../../utils/string";
+import {db} from "../../database/database";
+import {upsertUserInfo} from "../../database/userInfo";
+import {constants, text} from "../../providers/config";
+import {Command} from "../../structures/Command";
+import {format} from "../../utils/string";
 import pms from "pretty-ms";
-import { randRange, sampleArray } from "../../utils/utils";
+import {randRange, sampleArray} from "../../utils/utils";
+
 const cooldowns: Record<string, number> = {};
 
 export const command = new Command("work", "Gets you some money.")
@@ -21,7 +22,7 @@ export const command = new Command("work", "Gets you some money.")
 			);
 			return;
 		}
-		const info = await upsertUserInfo(int.user, int.guild?.id || '');
+		const info = await upsertUserInfo(int.user);
 		const obtained = randRange(...constants.work.amountRange);
 		cooldowns[int.user.id] = Date.now() + constants.work.cooldownMs;
 		await db.userInfo.update({ where: { id: info.id }, data: { balance: { increment: obtained } } });
