+++ src/commands/orders/worker/wash.ts
@@ -1,8 +1,7 @@
-import { CommandInteraction } from 'discord.js';
-import { PrismaClient } from '@prisma/client';
-import { permissions } from "../../../providers/permissions";
-import { Command } from "../../../structures/Command";
-import { getUserBalance, updateBalance } from "../../../database/userInfo";
+import {PrismaClient} from "@prisma/client";
+import {permissions} from "../../../providers/permissions";
+import {Command} from "../../../structures/Command";
+import {getUserBalance, updateBalance} from "../../../database/userInfo";
 
 const prisma = new PrismaClient();
 
@@ -19,7 +18,7 @@ export const command = new Command('wash', 'Wash a dish.')
     .addOption("boolean", (o) =>
         o.setName("inactive").setDescription("Include inactive orders too.")
     )
-    .setExecutor(async (interaction: CommandInteraction) => {
+    .setExecutor(async (interaction) => {
         const all = await prisma.dishes.findMany({ take: 10 });
         if (!all.length) {
             await interaction.reply("There are currently no dishes to wash.");
@@ -46,7 +45,7 @@ export const command = new Command('wash', 'Wash a dish.')
 
         setTimeout(async () => {
             const userInfo = await getUserBalance(interaction.user);
-            const donuts = userInfo.donuts + 60;
+            const donuts = userInfo.donuts ? userInfo.donuts + 60 : 60;
             await updateBalance(interaction.user, userInfo.balance, donuts);
 
             await interaction.followUp(`${interaction.user}, dish ${dish.id} has finished washing. You received 60 donuts.`);
