+++ src/events/interactionCreate.ts
@@ -1,19 +1,20 @@
-import { client } from "../providers/client";
-import { commandRegistry } from "../providers/commandManager";
-import { text } from "../providers/config";
-import { blacklist } from "../database/blacklist";
-import { StopCommandExecution } from "../utils/error";
+import {client} from "../providers/client";
+import {commandRegistry} from "../providers/commandManager";
+import {text} from "../providers/config";
+import {blacklist} from "../database/blacklist";
+import {StopCommandExecution} from "../utils/error";
+import {ChatInputCommandInteraction} from "discord.js";
 
 // Custom logging functions
 function logInteractionAlreadyReplied() {
 	console.log("The reply to this interaction has already been sent or deferred.");
 }
 
-function logBlacklistedUserInteraction(userId) {
+function logBlacklistedUserInteraction(userId: string) {
 	console.log(`Interaction failed due to user being blacklisted: ${userId}`);
 }
 
-function logException(message) {
+function logException(message: string) {
 	console.log(`Interaction failed with exception: ${message}`);
 }
 
@@ -36,12 +37,16 @@ client.on("interactionCreate", async (int) => {
 			if (!command) throw new Error(`Unregistered command ${int.commandName}`);
 			// TODO remove this and use discord builtin when permissions get better
 			for (const perm of command.permissions) await perm.check(int);
-			await command.executor(int);
+			// TODO we're assuming this is a ChatInputCommandInteraction, implement other types later.
+			if (!(int instanceof ChatInputCommandInteraction)) throw new Error("Unsupported interaction type, please use a chat input command.");
+			await command.executor(int as ChatInputCommandInteraction<"cached">);
 		}
 	} catch (e) {
 		if (!(e instanceof StopCommandExecution)) {
 			if (int.isCommand()) {
-				logException(e.message);
+				logException((e as typeof e & {
+					message: any
+				}).message); // pretend everything is fine and message exists...
 				int.reply({ content: text.errors.exception, ephemeral: true }).catch();
 			}
 			console.error(e);
